add tilemap,animation

+++ src/base64_decoder.cpp
@@ -32,21 +32,21 @@ char* decode(const char* code, int* filesize)
 	int loop = len / 4; if (flag > 0) loop -=1;
 	int i = 0;
 	for (; i < loop; ++i) {
-		char c0 = table[code[i * 4 + 0]]; assert(c0 >= 0);
-		char c1 = table[code[i * 4 + 1]]; assert(c1 >= 0);
-		char c2 = table[code[i * 4 + 2]]; assert(c2 >= 0);
-		char c3 = table[code[i * 4 + 3]]; assert(c3 >= 0);
+		char c0 = table[(int)code[i * 4 + 0]]; assert(c0 >= 0);
+		char c1 = table[(int)code[i * 4 + 1]]; assert(c1 >= 0);
+		char c2 = table[(int)code[i * 4 + 2]]; assert(c2 >= 0);
+		char c3 = table[(int)code[i * 4 + 3]]; assert(c3 >= 0);
 
 		file[i * 3 + 0] = ((c0 << 2) & 0xfc) | ((c1 >> 4) & 0x03);
 		file[i * 3 + 1] = ((c1 << 4) & 0xf0) | ((c2 >> 2) & 0x0f);
 		file[i * 3 + 2] = ((c2 << 6) & 0xc0) | ((c3 >> 0) & 0x3f);
 	}
 	if (flag > 0) {
-		char c0 = table[code[i * 4 + 0]]; assert(c0 >= 0);
-		char c1 = table[code[i * 4 + 1]]; assert(c1 >= 0);
+		char c0 = table[(int)code[i * 4 + 0]]; assert(c0 >= 0);
+		char c1 = table[(int)code[i * 4 + 1]]; assert(c1 >= 0);
 		file[i * 3 + 0] = ((c0 << 2) & 0xfc) | ((c1 >> 4) & 0x03);
 		if (flag < 2) {
-			char c2 = table[code[i * 4 + 2]]; assert(c2 >= 0);
+			char c2 = table[(int)code[i * 4 + 2]]; assert(c2 >= 0);
 			file[i * 3 + 1] = ((c1 << 4) & 0xf0) | ((c2 >> 2) & 0x0f);
 		}
 	}
